+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Implement CreateFork RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/497
+
 v0.64.0
 
 - Update vendored gitlab_git to b98c69470f52185117fcdb5e28096826b32363ca